Moved activity view
