+++ lib/presentation/router/router.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:movna/presentation/pages/home/home_page.dart';
+import 'package:movna/presentation/screens/activity/activity_screen.dart';
+import 'package:movna/presentation/screens/home/home_page.dart';
 
 part 'router.g.dart';
 
@@ -11,5 +12,15 @@ class HomeRoute extends GoRouteData {
   @override
   Widget build(BuildContext context, GoRouterState state) {
     return const HomePage();
+  }
+}
+
+@TypedGoRoute<ActivityRoute>(path: '/activity')
+class ActivityRoute extends GoRouteData {
+  const ActivityRoute();
+
+  @override
+  Widget build(BuildContext context, GoRouterState state) {
+    return const ActivityScreen();
   }
 }
